Fix: Compile fixes for RHEL 9.3 kernels
I'm not sure when some of these changes were backported, but the
unmodified lttng-modules compiles vs. 5.14.0-162.23.1

This disables the kmem and compaction modules.  I don't believe getting
these to compile will be easy, as the required struct declarations are
in vmlinux.h, and haven't been moved into mm/internal.h and mm/slab.h in
the RHEL sources.

Signed-off-by: Martin Hicks <martin@sr-research.com>